Code cleanup
Split up event route handlers:
  The even route handler file was getting quite large. I split it out
  into multiple files.

  I also moved the other routes into their own folders in preparation
  for similar changes

Delete dead features:
  The event-list code is something that was intended
  for supporting individual list curation. This idea may resurface, but in
  the meantime I am deleting it; so, we don't have to support it with
  tests and documentation.
