+++ api-server/routes/events/helpers/getIncludesForAdditionalEmbeddedChildModels.js
@@ -1,0 +1,9 @@
+module.exports = getIncludesForAdditionalEmbeddedChildModels
+
+function getIncludesForAdditionalEmbeddedChildModels(embeds, db) {
+  return embeds
+    .map(embedStr => {
+      return { model: db[embedStr] }
+    })
+    .filter(entry => entry !== null)
+}
