+++ api-server/routes/events/helpers/index.js
@@ -1,0 +1,9 @@
+const getIncludesForAdditionalEmbeddedChildModels = require('./getIncludesForAdditionalEmbeddedChildModels')
+const filterContactInfo = require('./filterContactInfo')
+const handleRequestForEventsBySingleTag = require('./handleRequestForEventsBySingleTag')
+
+module.exports = {
+  getIncludesForAdditionalEmbeddedChildModels,
+  filterContactInfo,
+  handleRequestForEventsBySingleTag
+}
